backend/requirements.txt
databases==0.9.0
fastapi==0.115.2
passlib==1.7.4
pydantic==2.9.2
python_jose==3.3.0
SQLAlchemy==2.0.35
uvicorn
python_dotenv==1.0.1
asyncpg==0.29.0
psycopg2-binary==2.9.10
requests==2.32.3
pillow==10.4.0
python-multipart==0.0.9
matplotlib==3.9.2
pytest_mock==3.14.0